+++ cassandra/io/asyncorereactor.py
@@ -43,13 +43,9 @@ log = logging.getLogger(__name__)
 
 _dispatcher_map = {}
 
-def _cleanup(loop_weakref):
-    try:
-        loop = loop_weakref()
-    except ReferenceError:
-        return
-
-    loop._cleanup()
+def _cleanup(loop):
+    if loop:
+        loop._cleanup()
 
 
 class WaitableTimer(Timer):
@@ -228,8 +224,6 @@ class AsyncoreLoop(object):
             dispatcher = _BusyWaitDispatcher()
         self._loop_dispatcher = dispatcher
 
-        atexit.register(partial(_cleanup, weakref.ref(self)))
-
     def maybe_start(self):
         should_start = False
         did_acquire = False
@@ -299,40 +293,43 @@ class AsyncoreLoop(object):
         log.debug("Dispatchers were closed")
 
 
+_global_loop = None
+atexit.register(partial(_cleanup, _global_loop))
+
+
 class AsyncoreConnection(Connection, asyncore.dispatcher):
     """
     An implementation of :class:`.Connection` that uses the ``asyncore``
     module in the Python standard library for its event loop.
     """
 
-    _loop = None
-
     _writable = False
     _readable = False
 
     @classmethod
     def initialize_reactor(cls):
-        if not cls._loop:
-            cls._loop = AsyncoreLoop()
+        global _global_loop
+        if not _global_loop:
+            _global_loop = AsyncoreLoop()
         else:
             current_pid = os.getpid()
-            if cls._loop._pid != current_pid:
+            if _global_loop._pid != current_pid:
                 log.debug("Detected fork, clearing and reinitializing reactor state")
                 cls.handle_fork()
-                cls._loop = AsyncoreLoop()
+                _global_loop = AsyncoreLoop()
 
     @classmethod
     def handle_fork(cls):
-        global _dispatcher_map
+        global _dispatcher_map, _global_loop
         _dispatcher_map = {}
-        if cls._loop:
-            cls._loop._cleanup()
-            cls._loop = None
+        if _global_loop:
+            _global_loop._cleanup()
+            _global_loop = None
 
     @classmethod
     def create_timer(cls, timeout, callback):
         timer = Timer(timeout, callback)
-        cls._loop.add_timer(timer)
+        _global_loop.add_timer(timer)
         return timer
 
     def __init__(self, *args, **kwargs):
@@ -344,14 +341,14 @@ class AsyncoreConnection(Connection, asyncore.dispatcher):
         self._connect_socket()
 
         # start the event loop if needed
-        self._loop.maybe_start()
+        _global_loop.maybe_start()
 
         init_handler = WaitableTimer(
             timeout=0,
             callback=partial(asyncore.dispatcher.__init__,
                              self, self._socket, _dispatcher_map)
         )
-        self._loop.add_timer(init_handler)
+        _global_loop.add_timer(init_handler)
         init_handler.wait(kwargs["connect_timeout"])
 
         self._writable = True
@@ -451,7 +448,7 @@ class AsyncoreConnection(Connection, asyncore.dispatcher):
         with self.deque_lock:
             self.deque.extend(chunks)
             self._writable = True
-        self._loop.wake_loop()
+        _global_loop.wake_loop()
 
     def writable(self):
         return self._writable